feat: refactor api directory
